Update Program.cs
Changed the movement algorithm such that aircraft does not need to have separate data for how it moves in each leg of the flight. The algorithm now calculates the distance required to travel to get to the next waypoint in the waypoint list and the angle, and on each tick, the velocity is decomposed into x and y components needed to reach the next waypoint.